refactor: last fixes
